extracted hook for parallax, moved sample items to separate ts file, removed developers.module.css for developers page, added color generation implementation for showcase section in LP.
